+++ src/app/components/homeParts/DailySummary.tsx
@@ -1,4 +1,4 @@
-import { MyContext } from "@/app/layout";
+import { MyContext } from "@/app/calendar/layout";
 import { Props, Transaction } from "@/types";
 import { financeCalculations } from "@/utils/financeCalculations";
 import {
