+++ src/app/components/homeParts/TransactionMenu.tsx
@@ -1,4 +1,4 @@
-import { MyContext } from "@/app/layout";
+import { MyContext } from "@/app/calendar/layout";
 import { Props } from "@/types";
 import {
   Box,
