+++ src/app/components/monthParts/CategoryChart.tsx
@@ -1,4 +1,4 @@
-import { MyContext } from "@/app/layout";
+import { MyContext } from "@/app/calendar/layout";
 import { Pie } from "react-chartjs-2";
 import {
   ExpenseCategory,
