+++ src/app/components/monthParts/MonthSelector.tsx
@@ -1,4 +1,4 @@
-import { MyContext } from "@/app/layout";
+import { MyContext } from "@/app/calendar/layout";
 import { Props } from "@/types";
 import { Box, Button } from "@mui/material";
 import { DatePicker, LocalizationProvider } from "@mui/x-date-pickers";
